get_patch_data: flags to filter by SHA and start count at 0
Filtering by SHA is interesting if you employ additional tools like PIT,
to find out which commits are relevant to your specific config. To do
this, create a file containing all SHAs to consider (line-by-line) and
supply it via the "-s" flag.

In general if you want to plot the total number of patches (by not
specifying a grep filter) you probably want to let them start at zero
which is now possible by supplying the "-0" flag.